Did Major Shift on `Goal`
This Library will be focused on pure Russian Roulette, it won't have inventory and Item
It was getting difficult for me to implement it structurally!

But I have surprise, and I'll reveal when I complete this library!

Other Changes:

Refactor inventory and weapon management: remove obsolete classes and interfaces, implement ShellInterface, LiveShell, and BlankShell, and enhance magazine and shotgun behaviors.