The Biggest and the Last Structural Change of this Repo - Shifting to `Stateless` Architecture
for months, I have been juggling between architecture and this final one and unexpectedly a **better one**

Refactor tons of things, Removed `State` and `Command` Pattern.

attrs shines! Heavily utilized it's `evolve` method for implement stateless architecture!

Introduced `processor` which has the job to process the task - like firing the shotgun, adding players and e.t.c!

Moved `Character's Behavioural` methods to `core` layer  - this made life easier!

This Change is Perfect, because I have understood what is a `library` ! it's job is to not do the job, but provide the tools and equipment to do the job! and that job is carried by the user who's going to use the library!
And `Stateless` architectuere shines, because `Stateful` architecture is normally use to build **Complete product**, which I was using it to build library!

Currently, the repo is formatted to Stateless - and this change was immediate - so it will take time adjust the whole repo according to it, currently, it's incomplete and has expected bugs!
**However** - things are clear to me and I tell you, this will be complete within this **Week**

Other changes (Generated through AI):

Refactor: Remove unused character states and interfaces

- Deleted MagazineState, JammedState, StockedState, PlayerBehaviour, AliveState, DeadState, PlayerState, ShotgunBehaviour, ShotgunState, LoadedState, UnLoadedState, and their respective interfaces.
- Removed the WeaponManager and RoundManager classes.
- Cleaned up the shotgun and magazine implementations, including their interfaces.
- Introduced a new central processor for handling game state transitions and actions.
- Updated game state management to reflect the new structure and removed deprecated methods.
- Enhanced the shell and magazine implementations to improve functionality and maintainability.